% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflist.R
\name{as.list.deflist}
\alias{as.list.deflist}
\title{Convert a deflist object to a list}
\usage{
\method{as.list}{deflist}(x, ...)
}
\arguments{
\item{x}{A deflist object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list containing the elements of the deflist object.
}
\description{
Convert a deflist object to a list
}
