% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflist.R
\name{length.deflist}
\alias{length.deflist}
\title{Retrieve the length of a deflist object}
\usage{
\method{length}{deflist}(x)
}
\arguments{
\item{x}{A deflist object.}
}
\value{
The length of the deflist object.
}
\description{
Retrieve the length of a deflist object
}
