% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflist.R
\name{[.deflist}
\alias{[.deflist}
\title{Subset a deflist object}
\usage{
\method{[}{deflist}(x, i)
}
\arguments{
\item{x}{A deflist object.}

\item{i}{Indices or names of the elements to be retrieved.}
}
\value{
A list containing the elements at the specified indices or names in the \code{deflist} object.
}
\description{
Subset a deflist object
}
