% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflist.R
\name{[[.deflist}
\alias{[[.deflist}
\title{Retrieve an element from a deflist object}
\usage{
\method{[[}{deflist}(x, i)
}
\arguments{
\item{x}{A deflist object.}

\item{i}{Index or name of the element to be retrieved.}
}
\value{
The element at the specified index or name in the deflist object.
}
\description{
Retrieve an element from a deflist object
}
