% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflist.R
\name{[[<-.deflist}
\alias{[[<-.deflist}
\title{Prevent assignment to an element in a deflist object}
\usage{
\method{[[}{deflist}(x, i) <- value
}
\arguments{
\item{x}{A deflist object.}

\item{i}{Index or name of the element to be assigned.}

\item{value}{Value to be assigned to the element.}
}
\value{
this function throws an error be design, no return value
}
\description{
Prevent assignment to an element in a deflist object
}
