% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflist.R
\name{[<-.deflist}
\alias{[<-.deflist}
\title{Prevent assignment to elements in a deflist object}
\usage{
\method{[}{deflist}(x, i) <- value
}
\arguments{
\item{x}{A deflist object.}

\item{i}{Indices or names of the elements to be assigned.}

\item{value}{Values to be assigned to the elements.}
}
\value{
this function throws an error be design, no return value
}
\description{
Prevent assignment to elements in a deflist object
}
